pyproject.toml
[tool.poetry]
name = "swde"
version = "0.1.0"
description = ""
authors = ["Gijs Hendriksen <info@gijshendriksen.nl>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.10.0"
lxml = "^4.8.0"
pandas = "^1.4.2"
tqdm = "^4.64.0"
transformers = "^4.18.0"
sentencepiece = "^0.1.96"
scikit-learn = "^1.0.2"
pytorch-lightning = "^1.6.0"
jupyterlab = "^3.3.3"
wandb = "^0.12.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
